Add restart connection to all api key read/write methods
