+++ server/__main__.py
@@ -6,7 +6,7 @@ import requests  # type: ignore
 from apscheduler.schedulers.background import BackgroundScheduler  # type: ignore
 
 from server.database.database_controller import remove_old_messages, set_message_retention_period  # type: ignore
-from server.database.connected_cars import clear_connected_cars  # type: ignore
+from server.database.cache import clear_connected_cars  # type: ignore
 from server.database.connection import set_db_connection  # type: ignore
 from server.database.time import timestamp  # type: ignore
 from server.fleetv2_http_api.__main__ import main as run_server  # type: ignore
