+++ server/database/models.py
@@ -1,0 +1,8 @@
+import dataclasses
+
+
+@dataclasses.dataclass
+class AdminDB:
+    id: int
+    name: str
+    key: str
