+++ server/database/security.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 from typing import Optional
-import dataclasses
 import random
 import string
 
@@ -8,27 +7,12 @@ from sqlalchemy import Engine, Select, func, select
 from sqlalchemy.orm import Session
 
 from server.database.connection import get_connection_source, AdminBase as _AdminBase
+from server.database.restart_connection import db_access_method as _db_access_method
+from server.database.cache import get_loaded_admins, store_admin
+from server.database.models import AdminDB
 
 
-_loaded_admins: list[AdminDB] = []
-
-
-def clear_loaded_admins() -> None:
-    global _loaded_admins
-    _loaded_admins.clear()
-
-
-def get_loaded_admins() -> list[AdminDB]:
-    return _loaded_admins.copy()
-
-
-@dataclasses.dataclass
-class AdminDB:
-    id: int
-    name: str
-    key: str
-
-
+@_db_access_method
 def add_admin_key(name: str, connection_source: Optional[Engine] = None) -> str:
     if connection_source is None:
         connection_source = get_connection_source()
@@ -46,20 +30,7 @@ def add_admin_key(name: str, connection_source: Optional[Engine] = None) -> str:
             return _admin_added_msg(name, key)
 
 
-def _admin_added_msg(name: str, key: str) -> str:
-    return f"Admin '{name}' added with key:\n\n{key}\n\n"
-
-
-def _admin_already_exists_msg(name: str) -> str:
-    return f"Admin with name '{name}' already exists."
-
-
-def _create_admin_table_if_it_does_not_exist(connection_source: Engine) -> None:
-    with connection_source.connect() as connection:
-        if not connection_source.dialect.has_table(connection, _AdminBase.__tablename__):
-            _AdminBase.metadata.create_all(connection_source)
-
-
+@_db_access_method
 def get_admin(key: str) -> AdminDB | None:
     loaded_admins = get_loaded_admins()
     for admin in loaded_admins:
@@ -79,10 +50,11 @@ def get_admin(key: str) -> AdminDB | None:
         else:
             admin_base: _AdminBase = result[0]
             admin = AdminDB(id=admin_base.id, name=admin_base.name, key=admin_base.key)
-            _loaded_admins.append(admin)
+            store_admin(admin)
             return admin
 
 
+@_db_access_method
 def admin_selection(key: str) -> Select:
     return select(_AdminBase).where(_AdminBase.key == key)
 
@@ -99,3 +71,17 @@ def number_of_admin_keys(connection: Optional[Engine] = None) -> int:
 
 def _generate_key() -> str:  # pragma: no cover
     return "".join(random.choice(string.ascii_letters) for _ in range(30))
+
+
+def _admin_added_msg(name: str, key: str) -> str:
+    return f"Admin '{name}' added with key:\n\n{key}\n\n"
+
+
+def _admin_already_exists_msg(name: str) -> str:
+    return f"Admin with name '{name}' already exists."
+
+
+def _create_admin_table_if_it_does_not_exist(connection_source: Engine) -> None:
+    with connection_source.connect() as connection:
+        if not connection_source.dialect.has_table(connection, _AdminBase.__tablename__):
+            _AdminBase.metadata.create_all(connection_source)
