+++ tests/database/test_security.py
@@ -5,10 +5,10 @@ sys.path.append(".")
 import unittest
 from unittest.mock import patch, Mock
 from server.database.connection import set_test_db_connection
+from server.database.cache import clear_loaded_admins
 from server.database.security import (
     get_admin,
     add_admin_key,
-    clear_loaded_admins,
     AdminDB,
     get_loaded_admins,
     number_of_admin_keys,
